Prompt user to upgrade URI before equipping
