+++ components/OldCharacterURIAlert.tsx
@@ -1,38 +1,17 @@
 import { Button, Text } from '@chakra-ui/react';
-import { useMemo } from 'react';
-import { useNetwork } from 'wagmi';
 
 import {
   PlayerActions,
   useCharacterActions,
 } from '@/contexts/CharacterActionsContext';
 import { useGame } from '@/contexts/GameContext';
-import { getChainLabelFromId } from '@/lib/web3';
-import { BASE_CHARACTER_URI } from '@/utils/constants';
 
 import { Alert } from './Alert';
 
 export const OldCharacterURIAlert: React.FC = () => {
   const { character } = useGame();
-  const { chain } = useNetwork();
-  const { openActionModal, selectCharacter } = useCharacterActions();
-
-  const uriNeedsUpgraded = useMemo(() => {
-    if (!(chain && character)) return false;
-    const chainLabel = getChainLabelFromId(chain.id);
-    const { uri } = character;
-    const potentialCID = uri
-      .split('/')
-      .filter(s => !!s)
-      .pop();
-
-    if (!(chainLabel && potentialCID)) return false;
-
-    const baseURI = uri.replace(potentialCID, '');
-    if (baseURI !== `${BASE_CHARACTER_URI}${chainLabel}/`) return false;
-
-    return potentialCID.match(/^[a-zA-Z0-9]{46,59}$/);
-  }, [chain, character]);
+  const { openActionModal, selectCharacter, uriNeedsUpgraded } =
+    useCharacterActions();
 
   if (!(character && uriNeedsUpgraded)) return null;
